move constants and analytics files, move ai-streaming module
