+++ enterprise/frontend/src/metabase-enterprise/metabot/components/Metabot.tsx
@@ -5,12 +5,12 @@ import { t } from "ttag";
 import MetabotFailure from "assets/img/metabot-failure.svg?component";
 import ErrorBoundary from "metabase/ErrorBoundary";
 import { useSelector } from "metabase/lib/redux";
+import { trackMetabotChatOpened } from "metabase/metabot/analytics";
 import { Sidebar } from "metabase/nav/containers/MainNavbar/MainNavbar.styled";
 import type { SuggestionModel } from "metabase/rich_text_editing/tiptap/extensions/shared/types";
 import { getUser } from "metabase/selectors/user";
 import { Box, Button, Flex, Text } from "metabase/ui";
 
-import { trackMetabotChatOpened } from "../analytics";
 import { useMetabotAgent } from "../hooks";
 import type { MetabotAgentId } from "../state";
