+++ enterprise/frontend/src/metabase-enterprise/metabot/components/MetabotAdmin/MetabotAdminPage.unit.spec.tsx
@@ -15,7 +15,7 @@ import { renderWithProviders, screen, waitFor } from "__support__/ui";
 import {
   FIXED_METABOT_ENTITY_IDS,
   FIXED_METABOT_IDS,
-} from "metabase-enterprise/metabot/constants";
+} from "metabase/metabot/constants";
 import { hasPremiumFeature } from "metabase-enterprise/settings";
 import type { MetabotId, RecentItem } from "metabase-types/api";
 import {
