+++ enterprise/frontend/src/metabase-enterprise/metabot/components/MetabotAdmin/MetabotAdminSuggestedPrompts.unit.spec.tsx
@@ -8,7 +8,7 @@ import {
   setupRemoveMetabotPromptSuggestionEndpoint,
 } from "__support__/server-mocks/metabot";
 import { renderWithProviders, screen, waitFor } from "__support__/ui";
-import { FIXED_METABOT_IDS } from "metabase-enterprise/metabot/constants";
+import { FIXED_METABOT_IDS } from "metabase/metabot/constants";
 import type { SuggestedMetabotPrompt } from "metabase-types/api";
 
 import { MetabotPromptSuggestionPane } from "./MetabotAdminSuggestedPrompts";
