+++ enterprise/frontend/src/metabase-enterprise/metabot/components/MetabotDataStudioButton.tsx
@@ -3,10 +3,10 @@ import { t } from "ttag";
 import { METAKEY } from "metabase/lib/browser";
 import { useSelector } from "metabase/lib/redux";
 import * as Urls from "metabase/lib/urls";
+import { trackMetabotChatOpened } from "metabase/metabot/analytics";
 import { getLocation } from "metabase/selectors/routing";
 import { ActionIcon, Icon, Tooltip } from "metabase/ui";
 
-import { trackMetabotChatOpened } from "../analytics";
 import { useMetabotAgent } from "../hooks";
 
 export const MetabotDataStudioButton = () => {
