+++ frontend/src/metabase/api/ai-streaming/index.ts
