+++ frontend/src/metabase/api/ai-streaming/process-stream.ts
