+++ frontend/src/metabase/api/ai-streaming/schemas.ts
