+++ frontend/src/metabase/api/ai-streaming/test-utils.ts
