+++ frontend/src/metabase/api/ai-streaming/types.ts
